fix setting BodyImage create fn in utils getCityImage() for getting random picture
